+++ src/app/js/admin/fields/TransformerListItem.js
@@ -8,7 +8,7 @@ import ActionDeleteIcon from 'material-ui/svg-icons/action/delete';
 import { Field, FieldArray, propTypes as reduxFormPropTypes } from 'redux-form';
 
 import { polyglot as polyglotPropTypes } from '../../propTypes';
-import { fromFields } from '../';
+import { fromFields } from '../selectors';
 import FormSelectField from '../../lib/FormSelectField';
 import TransformerArgList from './TransformerArgList';
